fix(lock): cover native Windows teardown denials
