+++ src/sidecar-lock.ts
@@ -1,38 +1,18 @@
 import fsSync from "node:fs";
-import fs from "node:fs/promises";
-import path from "node:path";
 import { sameFileIdentity } from "./file-identity.js";
 import {
-  readSidecarLockSnapshot,
-  relativeSidecarLockPath,
-  releaseSidecarReclaimGuard,
   removeSidecarLockIfUnchanged,
-  removeStaleSidecarLockIfAllowed,
-  serializeSidecarLockPayload,
   sidecarLockSnapshotMatches,
-  sidecarLockSnapshotStillPresent,
-  sidecarReclaimGuardExists,
-  tryAcquireSidecarReclaimGuard,
   type SidecarLockSnapshot,
-  type SidecarLockStaleSnapshot,
 } from "./sidecar-lock-reclaim.js";
-import { FsSafeError } from "./errors.js";
-import { createNativeExclusiveFile, type NativeFileHandle } from "./native-operations.js";
-import type { Root } from "./root-impl.js";
+import { acquireSidecarLock, type HeldSidecarLock } from "./sidecar-lock-acquire.js";
 import { createHeldSidecarLockHandle } from "./sidecar-lock-handle.js";
 import type {
   SidecarLockAcquireOptions,
   SidecarLockHandle,
   SidecarLockHeldEntry,
-  SidecarLockRetryOptions,
   WithSidecarLockOptions,
 } from "./sidecar-lock-types.js";
-import {
-  computeSidecarLockDelayMs,
-  defaultSidecarLockShouldReclaim,
-  isTransientLockFileDenial,
-  maxTransientLockDenials,
-} from "./sidecar-lock-policy.js";
 export type { SidecarLockStaleSnapshot } from "./sidecar-lock-reclaim.js";
 export type {
   SidecarLockAcquireOptions,
@@ -43,23 +23,9 @@ export type {
   SidecarLockStaleRecovery,
   WithSidecarLockOptions,
 } from "./sidecar-lock-types.js";
-type SidecarFileHandle = Pick<NativeFileHandle, "close" | "stat" | "writeFile">;
-type HeldLock = {
-  refCount: number;
-  reentrantOwner?: string;
-  handle: SidecarFileHandle;
-  lockPath: string;
-  snapshot: SidecarLockSnapshot;
-  acquiredAt: number;
-  metadata: Record<string, unknown>;
-  releasePromise?: Promise<void>;
-  lockRoot?: Root;
-  parsePayload?: (raw: string) => unknown;
-  compromiseTimer?: NodeJS.Timeout;
-};
 type SidecarLockManagerState = {
   cleanupRegistered: boolean;
-  held: Map<string, HeldLock>;
+  held: Map<string, HeldSidecarLock>;
   reclaimCleanupRegistered: boolean;
   reclaimGuards: Set<string>;
 };
@@ -137,17 +103,6 @@ function snapshotMatchesSync(lockPath: string, observed: SidecarLockSnapshot): b
   }
 }
 
-async function resolveNormalizedTargetPath(targetPath: string): Promise<string> {
-  const resolved = path.resolve(targetPath);
-  const dir = path.dirname(resolved);
-  await fs.mkdir(dir, { recursive: true });
-  try {
-    return path.join(await fs.realpath(dir), path.basename(resolved));
-  } catch {
-    return resolved;
-  }
-}
-
 function releaseAllReclaimGuardsSync(state: SidecarLockManagerState): void {
   for (const reclaimGuardPath of state.reclaimGuards) {
     try {
@@ -177,7 +132,7 @@ function releaseAllLocksSync(state: SidecarLockManagerState): void {
 async function releaseHeldLock(
   state: SidecarLockManagerState,
   normalizedTargetPath: string,
-  held: HeldLock,
+  held: HeldSidecarLock,
   options: { force?: boolean } = {},
 ): Promise<boolean> {
   const current = state.held.get(normalizedTargetPath);
@@ -216,7 +171,11 @@ async function releaseHeldLock(
   }
 }
 
-function handleForHeldLock(state: SidecarLockManagerState, normalizedTargetPath: string, held: HeldLock) {
+function handleForHeldLock(
+  state: SidecarLockManagerState,
+  normalizedTargetPath: string,
+  held: HeldSidecarLock,
+) {
   return createHeldSidecarLockHandle({
     normalizedTargetPath,
     held,
@@ -243,244 +202,15 @@ export function createSidecarLockManager(key: string) {
   async function acquire<TPayload extends Record<string, unknown>>(
     options: SidecarLockAcquireOptions<TPayload>,
   ): Promise<SidecarLockHandle> {
-    ensureExitCleanupRegistered();
-    const normalizedTargetPath = await resolveNormalizedTargetPath(options.targetPath);
-    const lockPath = options.lockPath ?? `${normalizedTargetPath}.lock`;
-    const held = state.held.get(normalizedTargetPath);
-    if (
-      held &&
-      options.reentrantOwner !== undefined &&
-      held.reentrantOwner !== undefined &&
-      options.reentrantOwner === held.reentrantOwner
-    ) {
-      held.refCount += 1;
-      return handleForHeldLock(state, normalizedTargetPath, held);
-    }
-
-    const startedAt = Date.now();
-    const retry = options.retry ?? {};
-    const maxRetries = options.timeoutMs === Number.POSITIVE_INFINITY ? undefined : retry.retries;
-    const reclaimGuardPath = `${lockPath}.reclaim`;
-    let ownsReclaimGuard = false;
-    let attempt = 0;
-    // Bounded so a genuine denial still surfaces as EPERM, not a lock timeout.
-    let transientDenials = 0;
-    const withinDenialBudget = (): boolean => ++transientDenials <= maxTransientLockDenials;
-    // Waiting can fail on the caller's own retry or deadline limits. Classifying
-    // a denial as contention must not cost them the original diagnosis, so hand
-    // the denial back when no further attempt can be scheduled.
-    const retryOrRethrowDenial = async (denial: unknown): Promise<void> => {
-      try {
-        await waitForRetry();
-      } catch (waitError) {
-        if ((waitError as NodeJS.ErrnoException).code === "file_lock_timeout") throw denial;
-        throw waitError;
-      }
-    };
-    const waitForRetry = async (): Promise<void> => {
-      const elapsed = Date.now() - startedAt;
-      if (
-        (options.timeoutMs !== undefined &&
-          options.timeoutMs !== Number.POSITIVE_INFINITY &&
-          elapsed >= options.timeoutMs) ||
-        (maxRetries !== undefined && attempt >= maxRetries)
-      ) {
-        throw Object.assign(new Error(`file lock timeout for ${normalizedTargetPath}`), {
-          code: "file_lock_timeout",
-          lockPath,
-          normalizedTargetPath,
-        });
-      }
-      const remaining =
-        options.timeoutMs === undefined || options.timeoutMs === Number.POSITIVE_INFINITY
-          ? Number.POSITIVE_INFINITY
-          : Math.max(0, options.timeoutMs - elapsed);
-      const delay = Math.min(computeSidecarLockDelayMs(retry, attempt), remaining);
-      attempt += 1;
-      await new Promise((resolve) => setTimeout(resolve, delay));
-    };
-    try {
-      while (true) {
-        if (!ownsReclaimGuard && (await sidecarReclaimGuardExists(reclaimGuardPath))) {
-          await waitForRetry();
-          continue;
-        }
-        let handle: SidecarFileHandle | null = null;
-        let lockFileCreateDenied = false;
-        try {
-          const payload = await options.payload();
-          const { raw, ownershipToken } = serializeSidecarLockPayload(payload);
-          if (options.lockRoot) {
-            const relativeLockPath = relativeSidecarLockPath(options.lockRoot, lockPath);
-            try {
-              await options.lockRoot.create(relativeLockPath, raw, { mkdir: true, mode: 0o600 });
-            } catch (error) {
-              if (error instanceof FsSafeError && error.code === "already-exists") {
-                throw Object.assign(new Error("sidecar lock exists"), { code: "EEXIST" });
-              }
-              throw error;
-            }
-            handle = (await options.lockRoot.open(relativeLockPath)).handle;
-          } else {
-            try {
-              handle =
-                (await createNativeExclusiveFile(lockPath, 0o600)) ?? (await fs.open(lockPath, "wx"));
-            } catch (createError) {
-              lockFileCreateDenied = isTransientLockFileDenial(createError, lockPath);
-              throw createError;
-            }
-            await handle.writeFile(raw, "utf8");
-          }
-          const snapshot = { raw, payload, stat: await handle.stat(), ownershipToken };
-          const createdHeld: HeldLock = {
-            refCount: 1,
-            reentrantOwner: options.reentrantOwner,
-            handle,
-            lockPath,
-            snapshot,
-            acquiredAt: Date.now(),
-            metadata: options.metadata ?? {},
-            lockRoot: options.lockRoot,
-            parsePayload: options.parsePayload,
-          };
-          state.held.set(normalizedTargetPath, createdHeld);
-          if (ownsReclaimGuard) {
-            try {
-              await releaseSidecarReclaimGuard(state.reclaimGuards, reclaimGuardPath);
-              ownsReclaimGuard = false;
-            } catch (err) {
-              await releaseHeldLock(state, normalizedTargetPath, createdHeld, { force: true });
-              throw err;
-            }
-          }
-          const returnedHandle = handleForHeldLock(state, normalizedTargetPath, createdHeld);
-          const interval = options.compromiseCheckIntervalMs;
-          if (options.onCompromised && interval !== undefined && interval > 0) {
-            createdHeld.compromiseTimer = setInterval(() => {
-              void returnedHandle.verifyStillHeld().then((stillHeld) => {
-                if (!stillHeld && createdHeld.compromiseTimer) {
-                  clearInterval(createdHeld.compromiseTimer);
-                  createdHeld.compromiseTimer = undefined;
-                  options.onCompromised?.({ lockPath, normalizedTargetPath });
-                }
-              });
-            }, interval);
-            createdHeld.compromiseTimer.unref();
-          }
-          return returnedHandle;
-        } catch (err) {
-          if (handle) {
-            const failedSnapshot: SidecarLockSnapshot = { payload: null };
-            try {
-              failedSnapshot.stat = await handle.stat();
-            } catch {
-              // Best-effort cleanup of a failed exclusive create.
-            }
-            const current = state.held.get(normalizedTargetPath);
-            if (current?.handle === handle) {
-              state.held.delete(normalizedTargetPath);
-            }
-            // If payload serialization/write fails, the file may be empty or
-            // partial JSON, so remove while our exclusive handle is still open.
-            if (!options.lockRoot) {
-              await fs.rm(lockPath, { force: true }).catch(() => undefined);
-            }
-            await handle.close().catch(() => undefined);
-            // Windows can refuse removing an open file; retry after close but
-            // only if the path still points at the file identity we created.
-            await removeSidecarLockIfUnchanged(lockPath, failedSnapshot, {
-              lockRoot: options.lockRoot,
-              parsePayload: options.parsePayload,
-            });
-          }
-          if (lockFileCreateDenied && withinDenialBudget()) {
-            await retryOrRethrowDenial(err);
-            continue;
-          }
-          if ((err as { code?: unknown }).code !== "EEXIST") {
-            throw err;
-          }
-          if (ownsReclaimGuard) {
-            await releaseSidecarReclaimGuard(state.reclaimGuards, reclaimGuardPath);
-            ownsReclaimGuard = false;
-            continue;
-          }
-          const nowMs = Date.now();
-          let snapshot: SidecarLockSnapshot | null;
-          try {
-            snapshot = await readSidecarLockSnapshot(lockPath, {
-              lockRoot: options.lockRoot,
-              parsePayload: options.parsePayload,
-            });
-          } catch (readErr) {
-            if (!isTransientLockFileDenial(readErr, lockPath) || !withinDenialBudget())
-              throw readErr;
-            await retryOrRethrowDenial(readErr);
-            continue;
-          }
-          if (!snapshot) {
-            continue;
-          }
-          if (state.held.has(normalizedTargetPath)) {
-            await waitForRetry();
-            continue;
-          }
-          const shouldReclaim = options.shouldReclaim ?? defaultSidecarLockShouldReclaim;
-          if (
-            await shouldReclaim({
-              lockPath,
-              normalizedTargetPath,
-              payload: snapshot?.payload ?? null,
-              staleMs: options.staleMs,
-              nowMs,
-              heldByThisProcess: state.held.has(normalizedTargetPath),
-            })
-          ) {
-            if (
-              !(await sidecarLockSnapshotStillPresent(lockPath, snapshot, {
-                lockRoot: options.lockRoot,
-                parsePayload: options.parsePayload,
-              }))
-            ) {
-              continue;
-            }
-            const staleRecovery = options.staleRecovery ?? "fail-closed";
-            if (staleRecovery === "remove-if-unchanged") {
-              if (!(await tryAcquireSidecarReclaimGuard(state.reclaimGuards, reclaimGuardPath))) {
-                await waitForRetry();
-                continue;
-              }
-              ownsReclaimGuard = true;
-              const removal = await removeStaleSidecarLockIfAllowed({
-                lockPath,
-                normalizedTargetPath,
-                snapshot,
-                shouldRemoveStaleLock: options.shouldRemoveStaleLock,
-                lockRoot: options.lockRoot,
-                parsePayload: options.parsePayload,
-              });
-              if (removal === "removed" || removal === "changed") {
-                continue;
-              }
-              await releaseSidecarReclaimGuard(state.reclaimGuards, reclaimGuardPath);
-              ownsReclaimGuard = false;
-            }
-            throw Object.assign(new Error(`file lock stale for ${normalizedTargetPath}`), {
-              code: "file_lock_stale",
-              lockPath,
-              normalizedTargetPath,
-            });
-          }
-          await waitForRetry();
-        }
-      }
-    } finally {
-      if (ownsReclaimGuard) {
-        await releaseSidecarReclaimGuard(state.reclaimGuards, reclaimGuardPath).catch(
-          () => undefined,
-        );
-      }
-    }
+    return await acquireSidecarLock(options, {
+      held: state.held,
+      reclaimGuards: state.reclaimGuards,
+      ensureExitCleanupRegistered,
+      handleForHeldLock: (normalizedTargetPath, held) =>
+        handleForHeldLock(state, normalizedTargetPath, held),
+      releaseHeldLock: async (normalizedTargetPath, held, releaseOptions) =>
+        await releaseHeldLock(state, normalizedTargetPath, held, releaseOptions),
+    });
   }
 
   async function withLock<T, TPayload extends Record<string, unknown>>(
